AttributeDraft to interface. #942
