+++ sphere-models/src/main/java/io/sphere/sdk/products/attributes/AttributeDraftImpl.java
@@ -1,0 +1,25 @@
+package io.sphere.sdk.products.attributes;
+
+
+import com.fasterxml.jackson.databind.JsonNode;
+import io.sphere.sdk.models.Base;
+
+final class AttributeDraftImpl extends Base implements AttributeDraft {
+    private final String name;
+    private final JsonNode value;
+
+    AttributeDraftImpl(final String name, final JsonNode value) {
+        this.name = name;
+        this.value = value;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public JsonNode getValue() {
+        return value;
+    }
+}
